moves helper functions to utils file
