+++ src/build_types/build_types.ts
@@ -1,12 +1,13 @@
 import { convertType } from './convert_type';
+import { toCamelCase, buildDescription} from './utils';
 
 export function buildTypes(spec: Record<string, any>): string {
-  const description = spec?.agent?.description.split('\n').map((line: string) => `// ${line}`);
+  const agentDesc = buildDescription(spec?.agent?.description)
   const agentEcs = convertAgentFields(spec?.agent?.fields);
   
   return (`// ECS types
 
-${description.join('\n')}
+${agentDesc}
 export interface Agent {
 ${agentEcs}
 }
@@ -23,8 +24,4 @@ export function convertAgentFields(agentFields: Record<string, any>): unknown {
     convertedNameTypes.push(`  ${toCamelCase(value.name)}?: ${convertType(value.type)};`)
   }
   return convertedNameTypes.join('\n');
-}
-
-export function toCamelCase(name:string): string {
-  return name.replace(/(\.|_)([a-z])/g, (g:string) => g[1].toUpperCase());
 }
