+++ src/build_types/utils.ts
@@ -1,0 +1,8 @@
+export function toCamelCase(name:string): string {
+  return name.replace(/(\.|_)([a-z])/g, (g:string) => g[1].toUpperCase());
+}
+
+export function buildDescription(desc: string): string {
+  return desc.split('\n').map((line: string) => `// ${line}`).join('\n');
+}
+
